Move notepad actions/reducers to slices module
