+++ src/actions/index.js
@@ -1,5 +1,6 @@
+export * as notepad from '../slices/notepad.js'
 export * as projectFiles from '../slices/project-files.js'
-export * as storage from './storage.js'
+export * as storage from '../slices/storage.js'
 
 export { default as activity } from './activity'
 export { default as api } from './api'
@@ -18,7 +19,6 @@ export { default as list } from './list'
 export { default as metadata } from './metadata'
 export { default as nav } from './nav'
 export { default as note } from './note'
-export { default as notepad } from './notepad'
 export { default as ontology } from './ontology'
 export { default as panel } from './panel'
 export { default as photo } from './photo'
