+++ src/browser/actions.js
@@ -1,4 +1,5 @@
-export * as storage from '../actions/storage.js'
+export * as notepad from '../slices/notepad.js'
+export * as storage from '../slices/storage.js'
 
 export { default as activity } from '../actions/activity'
 export { default as api } from '../actions/api'
@@ -10,7 +11,6 @@ export { default as item } from '../actions/item'
 export { default as list } from '../actions/list'
 export { default as metadata } from '../actions/metadata'
 export { default as note } from '../actions/note'
-export { default as notepad } from '../actions/notepad'
 export { default as ontology } from '../actions/ontology'
 export { default as photo } from '../actions/photo'
 export { default as project } from '../actions/project'
