+++ src/constants/index.js
@@ -19,7 +19,6 @@ export { default as METADATA } from './metadata'
 export { default as MIME } from './mime'
 export { default as NAV } from './nav'
 export { default as NOTE } from './note'
-export { default as NOTEPAD } from './notepad'
 export { default as ONTOLOGY } from './ontology'
 export { default as PANEL } from './panel'
 export { default as PHOTO } from './photo'
