+++ src/constants/note.js
@@ -1,5 +1,4 @@
 export default {
-  AUTOSAVE_DELAY: 5000,
   CREATE: 'note.create',
   EXPORT: 'note.export',
   INSERT: 'note.insert',
