--- src/constants/notepad.js
@@ -1,4 +1,0 @@
-export default {
-  RESTORE: 'notepad.restore',
-  UPDATE: 'notepad.update'
-}
